Transifex: modularize and migrate to python3
